requirements.txt
langgraph
langchain
langchain-openai
pandas
scikit-learn
scipy
python-dotenv
